+++ Changes.md
@@ -6,6 +6,7 @@ Breaking Changes
 
 - Dispatcher : Removed `createMatching()` method.
 - Process : Removed non-const variant of the `handleException()` method.
+- StringPlug : Removed deprecated `precomputedHash` argument from `getValue()` method.
 
 1.3.x.x (relative to 1.3.3.0)
 =======